+++ NEWS.md
@@ -11,6 +11,8 @@ the launching the app.
 was created with `fluidPage` or `navbarPage`: always create
 `ui.R`.
 - `send_message` also appends `-` to namespace in `ns2`.
+- Added input handlers in scaffold.
+- Fix update of endpoint utilities.
 
 # leprechaun 1.0.0.9000